requirement.txt
asgiref==3.5.2
backoff==2.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-widget-tweaks==1.4.12
idna==3.4
Pillow==9.3.0
pycryptodome==3.15.0
pyotp==2.7.0
pyrate-limiter==2.8.4
ratelimit==2.2.1
requests==2.28.1
requests-ratelimiter==0.4.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.13
